Changed SONGDETECTOR5.M to be compatible with Matlab R2014a, which appears to have more sensitive type requirements for its CONV.M function. I explicitly type casted the threshold crossing (logical) into double precision to be used for convolution.

+++ toolbox/songDetector5.m
@@ -47,7 +47,7 @@ powerSong = mean(abs(b(minNdx:maxNdx,:)), 1);
 powerNonSong = mean(abs(b([1:minNdx-1,maxNdx+1:end],:)), 1) + eps;
 songRatio = powerSong./powerNonSong;
 thresCross = (songRatio>ratioThreshold);
-songDet = conv(thresCross,windowAvg);
+songDet = conv(double(thresCross),windowAvg);
 tElapsed = toc;
 
 score = max(songDet);
